requirements.txt
streamlit
google-generativeai
Pillow
pytesseract
langdetect
gTTS
pandas
matplotlib
seaborn
python-dotenv
plotly
numpy
scikit-learn
